func_static update for platforms
solid states of func static with the platform key set now changes their solid contents when hide or show is executed. 
platform 1 now now sets a func static to allow bullets to pass through instead of being solid to all.
when hide is executed, the player can also pass
when show is executed, it reverts to only letting bullets pass.