Jump Through Platform activate and de-activate
Main code needed for Jump through platforms through use of a trigger above and below the platform.  Set the platform key to 1 for the func static and the platform_cant_hide key to 1 as well.  For the trigger_multi below set the jumpthrough key to 0.  For the trigger_multi pointing to the platform above, set it to jumpthrough 1.  Note that new entities have been added to simplify the process and less work for the mappers.