expand door trigger bounds key added
Mappers can now expand the default size of the touch trigger to open doors sooner.  The new default value is 200 the func.def file. This fixes the player having to keep stopping to wait for doors to open... HC is much faster than D3's gameplay.  It was made controllable just incase the mapper wants it default touch distance.  An example would be elevators and hidden doors.