Charge Attack system started
The basics of the player charge attack system are in.  Dashing charge attacks can be done on ground and air.  Offensive Dashing.  This includes allowing the player only being able to do x amount in a row, a counter to add charges back to maximum amount, and info needed for the hud to display how many times the player can currently charge.  All soft code has also been updated for this.  System still needs work for fx, animation and more.  Hard code still needs extra bug checks added in similar to those in for the slide move.  Player holds button_5 and presses forward to perform.