save/restore game crash fixes and hurtbox
2 issues in the player.cpp causing the game to crash when loading a save file have been fixed.
Added hurtbox event.  This works similar to the killbox (telefrag check).  The hurtbox is called from Enemies .  If they are over lapping the player, the player is damaged... thus triggering the knock back animation.  This fixes a bug where the player could overlap enemies if the invulnerability after damage state ended.