Developers Notes Added
Details info on the updates for the HardQore 2 port to Dhewm 3; "Hard Corps."
This includes how to set up jump through platforms, list of new keys for entities, and a full list of changes as they happen.  Finally a document detailing what files were modified in the sdk and what was done to them.