fixed noaim
previous update broke  no aim working for player bots....  the bots are used for many enemy soldier types in Hard Corps instead of Zsec.
when noaim is set to 1 in the enemy bot's def of map entity, it will only shoot straight ahead.  When set to 0 it is free to aim as normal.
This is a common feature in side scrollers.