friendly fire key check added
The projectile checks if it's owner is idAI and the new "friendly_fire" key is 1/0 in it's def file.  If set to 1, they can hurt other actors.  If set to 0, their projectiles will collide and hurt ONLY the player.  Another feature often found in 2d games.  Prevents enemies from killing each other when closely grouped.  Search for "rev 2021" in the code.