build 0.8.54 notes
this build adds a new sniper variant to the imp and random running animations for zombies so that they don't bunch up together in maps.  The big addition is the new key never_stop_moving for some enemies.  This was done because enemies would loose the player and completely give up chase when they shouldn't.  They would also not pursue the player when on a platform higher or lower than the player.  this would result in enemies just spawning in and standing around if the 2d level design was a bit complex.